ExamRAG/requirements.txt
pdfminer.six==20231228
langchain-community==0.2.19
faiss-cpu==1.8.0
tiktoken==0.7.0
python-dotenv==1.0.1
rich==13.7.1
streamlit==1.31.0


langchain==0.2.19         # core LC - matching community version
langchain-openai==0.1.8   # OpenAI wrapper (new SDK)
openai==1.25.0            # new SDK
